refactor: Main페이지 내 세부 로직 컴포넌트로 분리
